Move Material scrollbar regression test to material
